Cp 0904

+++ tests/queries/4_cnch_stateless/10016_test_bin.sql
@@ -3,7 +3,7 @@ SELECT bin(0);
 SELECT bin(42);
 select bin(18446744073709551615); -- 2^64 - 1
 select bin(9223372036854775808); -- 2^63
-SELECT bin(42.42);
+SELECT bin(42.42::Float64);
 SELECT bin('mooncake');
 SELECT bin(toDate('2020-08-20'));
 SELECT bin(toDateTime('2021-08-20 07:35:16'));
